Move `BoardLoader` into `loader` dir.
