+++ packages/breadboard/src/kits/load.ts
@@ -5,7 +5,7 @@
  */
 
 import { inspect } from "../inspector/index.js";
-import { BoardLoader, loadWithFetch } from "../loader.js";
+import { BoardLoader, loadWithFetch } from "../loader/loader.js";
 import { BoardRunner } from "../runner.js";
 import { GraphDescriptor, Kit, KitManifest, NodeHandler } from "../types.js";
 import { asRuntimeKit } from "./ctors.js";
